feat: Thêm file train + tunning của hai model
