+++ .gitignore
@@ -1,3 +1,3 @@
-data/
+data/Data-for-Data-Mining-Project - Source.csv
 Manh_data/
 __pycache__
